requirements.txt
discord.py @ git+https://github.com/Rapptz/discord.py
asyncio == 3.4.3 
requests == 2.28.2 
aiomysql == 0.1.1 
pillow == 8.4.0 
matplotlib == 3.8.2
pynacl == 1.5.0
topggpy == 1.4.0 
covid == 2.5.2 
covid-19 == 0.5.1 
jishaku == 2.5.1 
easy-pil == 0.1.9 
python-dotenv == 1.0.0 
qrcode == 7.4.2 
pytz == 2022.7.1
pyfiglet == 0.8.post1 
psutil == 5.9.4
pymsql == 1.1.1